[#7242] Consolidate duplicate irsync tests
There are several tests in test_icommands_file_operations which are
almost exactly identical to several tests in test_irsync. These have
been consolidated and updated appropriately to prevent failures. Some
other miscellaneous irsync tests have been moved out of
test_icommands_file_operations and into test_irsync for logical clarity.
