+++ scripts/irods/test/test_irsync.py
@@ -11,6 +11,7 @@ import shutil
 
 from .resource_suite import ResourceBase
 from .. import lib
+from .. import test
 from . import ustrings
 
 class Test_iRsync(ResourceBase, unittest.TestCase):
@@ -25,6 +26,20 @@ class Test_iRsync(ResourceBase, unittest.TestCase):
     def tearDown(self):
         shutil.rmtree(self.testing_tmp_dir)
         super(Test_iRsync, self).tearDown()
+
+    def iput_r_large_collection(self, user_session, base_name, file_count, file_size):
+        local_dir = os.path.join(self.testing_tmp_dir, base_name)
+        local_files = lib.make_large_local_tmp_dir(local_dir, file_count, file_size)
+        user_session.assert_icommand(['iput', '-r', local_dir])
+        rods_files = set(user_session.get_entries_in_collection(base_name))
+        self.assertTrue(set(local_files) == rods_files,
+                        msg="Files missing:\n" + str(set(local_files) - rods_files) + "\n\n" +
+                            "Extra files:\n" + str(rods_files - set(local_files)))
+        vault_files = set(os.listdir(os.path.join(user_session.get_vault_session_path(), base_name)))
+        self.assertTrue(set(local_files) == vault_files,
+                        msg="Files missing from vault:\n" + str(set(local_files) - vault_files) + "\n\n" +
+                            "Extra files in vault:\n" + str(vault_files - set(local_files)))
+        return (local_dir, local_files)
 
     def test_irsync(self):
         filepath = os.path.join(self.admin.local_session_dir, 'file')
@@ -77,6 +92,84 @@ class Test_iRsync(ResourceBase, unittest.TestCase):
 
         self.user0.environment_file_contents = user0_env_backup
 
+    def test_irsync_r_dir_to_coll(self):
+        base_name = "test_irsync_r_dir_to_coll"
+        local_dir = os.path.join(self.testing_tmp_dir, base_name)
+        file_names = set(lib.make_large_local_tmp_dir(local_dir, file_count=1000, file_size=100))
+
+        self.user0.assert_icommand("irsync -r " + local_dir + " i:" + base_name, "STDOUT_SINGLELINE", ustrings.recurse_ok_string())
+        self.user0.assert_icommand("ils", 'STDOUT_SINGLELINE', base_name)
+        rods_files = set(self.user0.get_entries_in_collection(base_name))
+        self.assertTrue(file_names == rods_files,
+                        msg="Files missing:\n" + str(file_names - rods_files) + "\n\n" +
+                            "Extra files:\n" + str(rods_files - file_names))
+
+        vault_files_post_irsync = set(os.listdir(os.path.join(self.user0.get_vault_session_path(),
+                                                              base_name)))
+
+        if not test.settings.RUN_IN_TOPOLOGY:
+            self.assertTrue(file_names == vault_files_post_irsync,
+                            msg="Files missing from vault:\n" + str(file_names - vault_files_post_irsync) + "\n\n" +
+                                "Extra files in vault:\n" + str(vault_files_post_irsync - file_names))
+
+    def test_irsync_r_coll_to_dir(self):
+        base_name_source = "test_irsync_r_coll_to_dir_source"
+        file_names = set(self.iput_r_large_collection(
+            self.user0, base_name_source, file_count=1000, file_size=100)[1])
+        local_dir = os.path.join(self.testing_tmp_dir, "test_irsync_r_coll_to_dir_target")
+        self.user0.assert_icommand("irsync -r i:" + base_name_source + " " + local_dir, "EMPTY")
+        self.user0.assert_icommand("ils", 'STDOUT_SINGLELINE', base_name_source)
+        rods_files_source = set(self.user0.get_entries_in_collection(base_name_source))
+        self.assertTrue(file_names == rods_files_source,
+                        msg="Files missing:\n" + str(file_names - rods_files_source) + "\n\n" +
+                            "Extra files:\n" + str(rods_files_source - file_names))
+
+        local_files = set(os.listdir(local_dir))
+        self.assertTrue(file_names == local_files,
+                        msg="Files missing from local dir:\n" + str(file_names - local_files) + "\n\n" +
+                            "Extra files in local dir:\n" + str(local_files - file_names))
+
+        if not test.settings.RUN_IN_TOPOLOGY:
+            vault_files_post_irsync_source = set(
+                os.listdir(os.path.join(self.user0.get_vault_session_path(), base_name_source)))
+
+            self.assertTrue(file_names == vault_files_post_irsync_source,
+                            msg="Files missing from vault:\n" + str(file_names - vault_files_post_irsync_source) + "\n\n" +
+                                "Extra files in vault:\n" + str(vault_files_post_irsync_source - file_names))
+
+    def test_irsync_r_coll_to_coll(self):
+        base_name_source = "test_irsync_r_coll_to_coll_source"
+        file_names = set(self.iput_r_large_collection(
+            self.user0, base_name_source, file_count=1000, file_size=100)[1])
+        base_name_target = "test_irsync_r_coll_to_coll_target"
+        self.user0.assert_icommand("irsync -r i:" + base_name_source + " i:" + base_name_target, "EMPTY")
+        self.user0.assert_icommand("ils", 'STDOUT_SINGLELINE', base_name_source)
+        self.user0.assert_icommand("ils", 'STDOUT_SINGLELINE', base_name_target)
+        rods_files_source = set(self.user0.get_entries_in_collection(base_name_source))
+        self.assertTrue(file_names == rods_files_source,
+                        msg="Files missing:\n" + str(file_names - rods_files_source) + "\n\n" +
+                            "Extra files:\n" + str(rods_files_source - file_names))
+
+        rods_files_target = set(self.user0.get_entries_in_collection(base_name_target))
+        self.assertTrue(file_names == rods_files_target,
+                        msg="Files missing:\n" + str(file_names - rods_files_target) + "\n\n" +
+                            "Extra files :\n" + str(rods_files_target - file_names))
+
+        if not test.settings.RUN_IN_TOPOLOGY:
+            vault_files_post_irsync_source = set(
+                os.listdir(os.path.join(self.user0.get_vault_session_path(), base_name_source)))
+
+            self.assertTrue(file_names == vault_files_post_irsync_source,
+                            msg="Files missing from vault:\n" + str(file_names - vault_files_post_irsync_source) + "\n\n" +
+                                "Extra files in vault:\n" + str(vault_files_post_irsync_source - file_names))
+
+            vault_files_post_irsync_target = set(
+                os.listdir(os.path.join(self.user0.get_vault_session_path(), base_name_target)))
+
+            self.assertTrue(file_names == vault_files_post_irsync_target,
+                            msg="Files missing from vault:\n" + str(file_names - vault_files_post_irsync_target) + "\n\n" +
+                                "Extra files in vault:\n" + str(vault_files_post_irsync_target - file_names))
+
     def test_irsync_r_nested_dir_to_coll(self):
         # test settings
         depth = 10
@@ -106,6 +199,13 @@ class Test_iRsync(ResourceBase, unittest.TestCase):
                             msg="Files missing:\n" + str(local_files - rods_files) + "\n\n" +
                             "Extra files:\n" + str(rods_files - local_files))
 
+            if not test.settings.RUN_IN_TOPOLOGY:
+                # compare local files with files in vault
+                files_in_vault = set(lib.files_in_dir(os.path.join(self.user0.get_vault_session_path(), partial_path)))
+                self.assertTrue(local_files == files_in_vault,
+                            msg="Files missing from vault:\n" + str(local_files - files_in_vault) + "\n\n" +
+                                "Extra files in vault:\n" + str(files_in_vault - local_files))
+
     def test_irsync_r_nested_dir_to_coll_large_files(self):
         # test settings
         depth = 4
@@ -134,6 +234,13 @@ class Test_iRsync(ResourceBase, unittest.TestCase):
             self.assertTrue(local_files == rods_files,
                             msg="Files missing:\n" + str(local_files - rods_files) + "\n\n" +
                             "Extra files:\n" + str(rods_files - local_files))
+
+            if not test.settings.RUN_IN_TOPOLOGY:
+                # compare local files with files in vault
+                files_in_vault = set(lib.files_in_dir(os.path.join(self.user0.get_vault_session_path(), partial_path)))
+                self.assertTrue(local_files == files_in_vault,
+                                msg="Files missing from vault:\n" + str(local_files - files_in_vault) + "\n\n" +
+                                "Extra files in vault:\n" + str(files_in_vault - local_files))
 
     def test_irsync_r_nested_coll_to_coll(self):
         # test settings
@@ -169,6 +276,14 @@ class Test_iRsync(ResourceBase, unittest.TestCase):
                             msg="Files missing:\n" + str(local_files - rods_files) + "\n\n" +
                             "Extra files:\n" + str(rods_files - local_files))
 
+            if not test.settings.RUN_IN_TOPOLOGY:
+                # compare local files with files in vault
+                files_in_vault = set(
+                    lib.files_in_dir(os.path.join(self.user0.get_vault_session_path(), source_partial_path)))
+                self.assertTrue(local_files == files_in_vault,
+                                msg="Files missing from vault:\n" + str(local_files - files_in_vault) + "\n\n" +
+                                    "Extra files in vault:\n" + str(files_in_vault - local_files))
+
             # now the same thing with dest subcollection
             self.user0.assert_icommand(['ils', dest_partial_path], 'STDOUT_SINGLELINE')
             ils_out = self.user0.get_entries_in_collection(dest_partial_path)
@@ -179,6 +294,14 @@ class Test_iRsync(ResourceBase, unittest.TestCase):
             self.assertTrue(local_files == rods_files,
                             msg="Files missing:\n" + str(local_files - rods_files) + "\n\n" +
                             "Extra files:\n" + str(rods_files - local_files))
+
+            if not test.settings.RUN_IN_TOPOLOGY:
+                # compare local files with files in vault
+                files_in_vault = set(
+                    lib.files_in_dir(os.path.join(self.user0.get_vault_session_path(), dest_partial_path)))
+                self.assertTrue(local_files == files_in_vault,
+                                msg="Files missing from vault:\n" + str(local_files - files_in_vault) + "\n\n" +
+                                "Extra files in vault:\n" + str(files_in_vault - local_files))
 
     def test_irsync_r_nested_coll_to_coll_large_files(self):
         # test settings
